add skeleton of flex-archspec
Note that we need to better understand the structure of the graph,
and how this maps to the JobSpec (the request for a machine).
I have not been able to get a match yet. I would also like to
add in compilers when that is done, which is something I
just added support for in archspec-go

Signed-off-by: vsoch <vsoch@users.noreply.github.com>